perf: adding types with jsdoc, refactoring and splitting files into functions
